renamed files for better continuity and understanding
