+++ pages/index.js
@@ -3,7 +3,7 @@ import { useEffect, useContext, useState } from 'react';
 import { appContext } from "./_app";
 import Layout from './components/Layout/Layout';
 import styles from '@/styles/Home.module.css';
-import HomePage from './Home/Home.js';
+import HomePage from './Home/index.js';
  
 export default function Home() {
 
@@ -14,7 +14,7 @@ export default function Home() {
   return (
     <>
       <Head>
-        <title>Deployed Branch</title>
+        <title>Development Branch</title>
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
