+++ styles/PreGamePage.module.css
@@ -5,16 +5,29 @@
     width: 100%;
     align-items: center;
     justify-content: center;
-    border: 1px solid red;
+    /* border: 1px solid red; */
 }
 
 .preGameContainer > * {
     flex: 1;
+    width: 100%;
 }
 
 .preGameTopSection {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
 
+}
+.preGameTopSection > h1 {
+    display: flex;
+    justify-content: center;
 }
 .preGameBottomSection {
 
+}
+
+.playerContainer {
+    border: 1px solid red;
+    padding: 5vh 5vh 5vh 5vh
 }
